+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Fixed a bug where TODOs were not being reported if they were located after a
+  multi-line comment with no TODOs in the same file
+  ([#1520](https://github.com/ianlewis/todos/issues/1520))
+
 ## [0.9.0-rc.2] - 2024-07-17
 
 ### Added in 0.9.0-rc.2